fix: community mapping

+++ src/main/java/ju/example/training_management_system/repository/TokenRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface TokenRepository extends JpaRepository<PasswordResetToken, Integer> {
+public interface TokenRepository extends JpaRepository<PasswordResetToken, Long> {
     PasswordResetToken findByToken(String token);
 
     void deleteByUserId(Long id);
